create structure to users
