+++ pkg/db/config.go
@@ -1,13 +1,35 @@
 package db
 
-import "database/sql"
+import (
+	"context"
+	"time"
+)
 
-const dbDriver = "postgres"
+type Config struct {
+	ID        int64
+	Key       string
+	Value     string
+	CreatedAt time.Time
+	UpdatedAt time.Time
+}
 
-func NewConnection(datasource string) (*sql.DB, error) {
-	conn, err := sql.Open(dbDriver, datasource)
-	if err != nil {
-		return nil, err
-	}
-	return conn, err
+type GetConfigByKeyParams struct {
+	Key string
+}
+
+const getConfigByKeyQuery = `
+SELECT id, key, value
+FROM configs
+WHERE key = $1
+`
+
+func (q *Queries) GetConfigByKey(ctx context.Context, arg GetConfigByKeyParams) (Config, error) {
+	row := q.db.QueryRowContext(ctx, getConfigByKeyQuery, arg.Key)
+	var c Config
+	err := row.Scan(
+		&c.ID,
+		&c.Key,
+		&c.Value,
+	)
+	return c, err
 }
