--- pkg/db/models.go
@@ -1,11 +1,0 @@
-package db
-
-import "time"
-
-type Config struct {
-	ID        int64
-	Key       string
-	Value     string
-	CreatedAt time.Time
-	UpdatedAt time.Time
-}
