+++ pkg/domain/user/handlers.go
@@ -1,0 +1,1 @@
+package user
